refactor: refactoring project using mods instead of declaring modules at root
