+++ Cargo.toml
@@ -10,6 +10,8 @@ rust-version = "1.63"
 version = "0.1.0"
 
 [dependencies]
+actix = "0.13"
+actix-rt = "2.2"
 abscissa_core = { version = "0.6.0" }
 abscissa_tokio = { version = "0.6.0-pre.2", features = ["actix"] }
 bip39 = "1"
