+++ src/messages/mod.rs
@@ -1,0 +1,3 @@
+//! Holds actors messages
+
+pub mod request;
